requirements.txt
google-search-results
python-dotenv
requests
beautifulsoup4
pandas